add fd fct and clean includes

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: tsiguenz <tsiguenz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/17 15:11:38 by tsiguenz          #+#    #+#             */
-/*   Updated: 2021/11/23 16:22:50 by tsiguenz         ###   ########.fr       */
+/*   Updated: 2021/11/24 18:22:04 by tsiguenz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 void	*ft_memchr(const void *s, int c, size_t n)
 {
-	unsigned int	i;
-	char			*p_s;
+	size_t	i;
+	char	*p_s;
 
 	i = 0;
 	p_s = (char *) s;
