+++ ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: tsiguenz <tsiguenz@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/17 16:08:13 by tsiguenz          #+#    #+#             */
-/*   Updated: 2021/11/23 16:12:20 by tsiguenz         ###   ########.fr       */
+/*   Updated: 2021/11/24 18:21:46 by tsiguenz         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	unsigned int	i;
-	char			*p_s1;
-	char			*p_s2;
+	size_t	i;
+	char	*p_s1;
+	char	*p_s2;
 
 	i = 0;
 	p_s1 = (char *) s1;
